refactor(prototyper): move the hart cells into the trap-stack base
HsmCell, RFenceCell, and the rfence Fifo are primitives that live in
the per-hart slot — they belong with the slot's owning module, not in
the business files. They move verbatim into trap_stack (module doc now
names the raw state), with the full-queue retry loops staying in
rfence.rs as same-crate impl blocks (the loop coordinates with the
business-side IPI handler). hsm/rfence shed their cell unsafe and
lose nothing else; the remaining CSR asm sites are the CSR-packaging
step's scope.

Boot logs byte-identical except three PMP region-address lines (image
size shifted; addresses move with it).
