+++ prototyper/prototyper/src/sbi/hsm.rs
@@ -1,184 +1,13 @@
-//! HSM extension; the per-hart cell views (`local_hsm`/`remote_hsm`,
-//! re-exported below) come from the safe accessors in [`super::trap_stack`].
+//! SBI HSM (Hart State Management) extension.
 
-use core::{
-    cell::UnsafeCell,
-    hint::spin_loop,
-    sync::atomic::{AtomicUsize, Ordering},
-};
 use riscv::register::mstatus::MPP;
-use rustsbi::{SbiRet, spec::hsm::hart_state};
+use rustsbi::SbiRet;
 
 use crate::riscv::current_hartid;
 use crate::sbi::hart_context::NextStage;
 
 use super::trap::boot::boot;
-use super::trap_stack::{hart_local, reset_hart};
-
-/// Special state indicating a hart is in the process of starting.
-const HART_STATE_START_PENDING_EXT: usize = usize::MAX;
-
-type HsmState = AtomicUsize;
-
-/// Cell for managing hart state and shared data between harts.
-pub(crate) struct HsmCell<T> {
-    status: HsmState,
-    inner: UnsafeCell<Option<T>>,
-}
-
-impl<T> HsmCell<T> {
-    /// Creates a new HsmCell with STOPPED state and no inner data.
-    pub const fn new() -> Self {
-        Self {
-            status: HsmState::new(hart_state::STOPPED),
-            inner: UnsafeCell::new(None),
-        }
-    }
-
-    /// Gets a local view of this cell for the current hart.
-    ///
-    /// # Safety
-    ///
-    /// Caller must ensure this cell belongs to the current hart.
-    #[inline]
-    pub unsafe fn local(&self) -> LocalHsmCell<'_, T> {
-        LocalHsmCell(self)
-    }
-
-    /// Gets a remote view of this cell for accessing from other harts.
-    #[inline]
-    pub fn remote(&self) -> RemoteHsmCell<'_, T> {
-        RemoteHsmCell(self)
-    }
-}
-
-/// View of HsmCell for operations on the current hart.
-pub struct LocalHsmCell<'a, T>(&'a HsmCell<T>);
-
-/// View of HsmCell for operations from other harts.
-pub struct RemoteHsmCell<'a, T>(&'a HsmCell<T>);
-
-// Mark HsmCell as safe to share between threads
-unsafe impl<T: Send> Sync for HsmCell<T> {}
-unsafe impl<T: Send> Send for HsmCell<T> {}
-
-impl<T> LocalHsmCell<'_, T> {
-    /// Attempts to transition hart from START_PENDING to STARTED state.
-    ///
-    /// Returns inner data if successful, otherwise returns current state.
-    #[inline]
-    pub fn start(&self) -> Result<T, usize> {
-        loop {
-            match self.0.status.compare_exchange(
-                hart_state::START_PENDING,
-                hart_state::STARTED,
-                Ordering::AcqRel,
-                Ordering::Relaxed,
-            ) {
-                Ok(_) => break Ok(unsafe { (*self.0.inner.get()).take().unwrap() }),
-                Err(HART_STATE_START_PENDING_EXT) => spin_loop(),
-                Err(s) => break Err(s),
-            }
-        }
-    }
-
-    /// Transitions hart to STOPPED state.
-    #[allow(unused)]
-    #[inline]
-    pub fn stop(&self) {
-        self.0.status.store(hart_state::STOPPED, Ordering::Release)
-    }
-
-    /// Transitions hart to SUSPENDED state.
-    #[allow(unused)]
-    #[inline]
-    pub fn suspend(&self) {
-        self.0
-            .status
-            .store(hart_state::SUSPENDED, Ordering::Relaxed)
-    }
-
-    /// Transitions hart to STARTED state.
-    #[allow(unused)]
-    #[inline]
-    pub fn resume(&self) {
-        self.0.status.store(hart_state::STARTED, Ordering::Relaxed)
-    }
-}
-
-impl<T: core::fmt::Debug> RemoteHsmCell<'_, T> {
-    /// Attempts to start a stopped hart by providing startup data.
-    ///
-    /// Returns true if successful, false if hart was not in STOPPED state.
-    #[inline]
-    pub fn start(&self, t: T) -> bool {
-        if self
-            .0
-            .status
-            .compare_exchange(
-                hart_state::STOPPED,
-                HART_STATE_START_PENDING_EXT,
-                Ordering::Acquire,
-                Ordering::Relaxed,
-            )
-            .is_ok()
-        {
-            unsafe { *self.0.inner.get() = Some(t) };
-            self.0
-                .status
-                .store(hart_state::START_PENDING, Ordering::Release);
-            true
-        } else {
-            false
-        }
-    }
-
-    /// Attempts to resume a suspended hart by providing resume data.
-    ///
-    /// Returns true if successful, false if hart was not in SUSPENDED state.
-    #[inline]
-    pub fn resume(&self, t: T) -> bool {
-        if self
-            .0
-            .status
-            .compare_exchange(
-                hart_state::SUSPENDED,
-                HART_STATE_START_PENDING_EXT,
-                Ordering::Acquire,
-                Ordering::Relaxed,
-            )
-            .is_ok()
-        {
-            unsafe { *self.0.inner.get() = Some(t) };
-            self.0
-                .status
-                .store(hart_state::START_PENDING, Ordering::Release);
-            true
-        } else {
-            false
-        }
-    }
-
-    /// Gets the current state of the hart.
-    #[allow(unused)]
-    #[inline]
-    pub fn get_status(&self) -> usize {
-        match self.0.status.load(Ordering::Relaxed) {
-            HART_STATE_START_PENDING_EXT => hart_state::START_PENDING,
-            normal => normal,
-        }
-    }
-
-    /// Checks if hart can receive IPIs (must be STARTED or SUSPENDED).
-    #[allow(unused)]
-    #[inline]
-    pub fn allow_ipi(&self) -> bool {
-        matches!(
-            self.0.status.load(Ordering::Relaxed),
-            hart_state::STARTED | hart_state::SUSPENDED
-        )
-    }
-}
+use super::trap_stack::{RemoteHsmCell, hart_local, reset_hart};
 
 /// Gets the local HSM cell for the current hart.
 pub(crate) use super::trap_stack::local_hsm;
